jadx-0.7.2 - B2 v4.8.10 (1304)

+++ sources/org/telegram/messenger/exoplayer2/source/chunk/MediaChunk.java
@@ -6,17 +6,17 @@ import org.telegram.messenger.exoplayer2.upstream.DataSpec;
 import org.telegram.messenger.exoplayer2.util.Assertions;
 
 public abstract class MediaChunk extends Chunk {
-    public final int chunkIndex;
+    public final long chunkIndex;
 
     public abstract boolean isLoadCompleted();
 
-    public MediaChunk(DataSource dataSource, DataSpec dataSpec, Format trackFormat, int trackSelectionReason, Object trackSelectionData, long startTimeUs, long endTimeUs, int chunkIndex) {
+    public MediaChunk(DataSource dataSource, DataSpec dataSpec, Format trackFormat, int trackSelectionReason, Object trackSelectionData, long startTimeUs, long endTimeUs, long chunkIndex) {
         super(dataSource, dataSpec, 1, trackFormat, trackSelectionReason, trackSelectionData, startTimeUs, endTimeUs);
         Assertions.checkNotNull(trackFormat);
         this.chunkIndex = chunkIndex;
     }
 
-    public int getNextChunkIndex() {
+    public long getNextChunkIndex() {
         return this.chunkIndex + 1;
     }
 }
